extract smaller components
